separating tests for cleaning from tests for loading and saving
